Join oracle pool with oracle service
