+++ src/neo/Oracle/OracleService.cs
@@ -1,5 +1,6 @@
 using Akka.Actor;
 using Akka.Configuration;
+using Neo.Cryptography.ECC;
 using Neo.IO;
 using Neo.IO.Actors;
 using Neo.Ledger;
@@ -11,6 +12,7 @@ using Neo.SmartContract.Native;
 using Neo.VM;
 using Neo.Wallets;
 using System;
+using System.Collections;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Linq;
@@ -22,31 +24,123 @@ namespace Neo.Oracle
 {
     public class OracleService : UntypedActor
     {
+        #region Sub classes
+
         internal class StartMessage { }
+
+        private class RequestItem : PoolItem
+        {
+            // Request
+
+            public readonly Transaction RequestTransaction;
+
+            // Response
+
+            public readonly Contract Contract;
+            public Transaction ResponseTransaction;
+            private ContractParametersContext ResponseContext;
+
+            public bool IsCompleted => ResponseContext?.Completed == true;
+
+            internal RequestItem(Contract contract, Transaction requestTx) : base(requestTx)
+            {
+                Contract = contract;
+                RequestTransaction = requestTx;
+            }
+
+            public bool AddSinature(ResponseItem response)
+            {
+                if (response.Data.TransactionRequestHash != RequestTransaction.Hash)
+                {
+                    return false;
+                }
+
+                if (ResponseTransaction == null)
+                {
+                    if (response.ResponseTx == null)
+                    {
+                        return false;
+                    }
+
+                    // Oracle service could attach the real TX
+
+                    ResponseTransaction = response.ResponseTx;
+                    ResponseContext = new ContractParametersContext(response.ResponseTx);
+                }
+
+                // TODO: Check duplicate call
+
+                return ResponseContext.AddSignature(Contract, response.OraclePub, response.Data.Signature) == true;
+            }
+        }
+
+        private class ResponseCollection : IEnumerable<ResponseItem>
+        {
+            public readonly DateTime Timestamp;
+            public readonly IDictionary<UInt160, ResponseItem> Items = new ConcurrentDictionary<UInt160, ResponseItem>();
+
+            public ResponseCollection(ResponseItem item)
+            {
+                Timestamp = item.Timestamp;
+                Add(item);
+            }
+
+            public bool Add(ResponseItem item)
+            {
+                return Items.TryAdd(item.Hash, item);
+            }
+
+            public IEnumerator<ResponseItem> GetEnumerator()
+            {
+                return (IEnumerator<ResponseItem>)Items.Values.ToArray().GetEnumerator();
+            }
+
+            IEnumerator IEnumerable.GetEnumerator()
+            {
+                return this.GetEnumerator();
+            }
+        }
+
+        private class ResponseItem : PoolItem
+        {
+            public readonly OraclePayload Msg;
+            public readonly OracleResponseSignature Data;
+            public readonly Transaction ResponseTx;
+
+            public ECPoint OraclePub => Msg.OraclePub;
+            public UInt160 Hash => Msg.Hash;
+
+            internal ResponseItem(OraclePayload payload, Transaction responseTx) : base(null)
+            {
+                Data = payload.OracleSignature;
+                ResponseTx = responseTx;
+            }
+        }
+
+        #endregion
 
         private const long MaxGasFilter = 1_000_000;
 
-        /// <summary>
-        /// A flag for know if the pool was ordered or not
-        /// </summary>
-        private bool _isDirty = false;
         private long _isStarted = 0;
-        private CancellationTokenSource _cancel;
+        private Contract _lastContract;
         private readonly IActorRef _localNode;
         private readonly KeyPair[] _accounts;
+        private CancellationTokenSource _cancel;
 
         /// <summary>
         /// Number of threads for processing the oracle
         /// </summary>
         private readonly Task[] _oracleTasks = new Task[4]; // TODO: Set this
+
         /// <summary>
         /// _oracleTasks will consume from this pool
         /// </summary>
         private readonly BlockingCollection<Transaction> _asyncPool = new BlockingCollection<Transaction>();
+
         /// <summary>
-        /// Sortable pool
+        /// Queue
         /// </summary>
-        private readonly List<Transaction> _orderedPool = new List<Transaction>();
+        private readonly SortedConcurrentDictionary<UInt256, Transaction> _queue;
 
         #region Protocols
 
@@ -63,9 +157,34 @@ namespace Neo.Oracle
         public Func<OracleRequest, OracleResponse> Oracle { get; }
 
         /// <summary>
-        /// Oracle pool
+        /// Pending user Transactions
         /// </summary>
-        public OraclePool Pool { get; }
+        private readonly SortedConcurrentDictionary<UInt256, RequestItem> _pendingOracleRequest;
+
+        /// <summary>
+        /// Pending oracle response Transactions
+        /// </summary>
+        private readonly SortedConcurrentDictionary<UInt256, ResponseCollection> _pendingOracleResponses;
+
+        /// <summary>
+        /// Total maximum capacity of transactions the pool can hold.
+        /// </summary>
+        public int PendingCapacity { get; }
+
+        /// <summary>
+        /// MemoryPool
+        /// </summary>
+        public MemoryPool MemPool { get; }
+
+        /// <summary>
+        /// Total requests in the pool.
+        /// </summary>
+        public int PendingRequestCount => _pendingOracleRequest.Count;
+
+        /// <summary>
+        /// Total responses in the pool.
+        /// </summary>
+        public int PendingResponseCount => _pendingOracleResponses.Count;
 
         /// <summary>
         /// Is started
@@ -80,6 +199,8 @@ namespace Neo.Oracle
         public OracleService(IActorRef localNode, Wallet wallet)
         {
             Oracle = Process;
+            MemPool = Blockchain.Singleton.MemPool;
+            PendingCapacity = MemPool.Capacity;
             _localNode = localNode;
 
             // Find oracle account
@@ -98,7 +219,21 @@ namespace Neo.Oracle
                 throw new ArgumentException(nameof(wallet));
             }
 
-            Pool = new OraclePool(Blockchain.Singleton.MemPool, Self, localNode, 30_000);
+            _queue = new SortedConcurrentDictionary<UInt256, Transaction>
+                (
+                Comparer<KeyValuePair<UInt256, Transaction>>.Create(SortEnqueuedRequest), PendingCapacity
+                );
+
+            // Create internal collections for pending request/responses
+
+            _pendingOracleRequest = new SortedConcurrentDictionary<UInt256, RequestItem>
+                (
+                Comparer<KeyValuePair<UInt256, RequestItem>>.Create(SortRequest), PendingCapacity
+                );
+            _pendingOracleResponses = new SortedConcurrentDictionary<UInt256, ResponseCollection>
+                (
+                Comparer<KeyValuePair<UInt256, ResponseCollection>>.Create(SortResponse), PendingCapacity
+                );
         }
 
         /// <summary>
@@ -108,29 +243,66 @@ namespace Neo.Oracle
         {
             switch (message)
             {
-                case StartMessage start:
+                case StartMessage _:
                     {
                         Start();
                         break;
                     }
                 case Transaction tx:
                     {
-                        // TODO: Check that it's a oracleTx
-                        // TODO: Should we check the max, or use MemoryPool?
+                        using var snapshot = Blockchain.Singleton.GetSnapshot();
+                        var contract = NativeContract.Oracle.GetOracleMultiSigContract(snapshot);
 
-                        lock (_orderedPool)
+                        switch (tx.Version)
                         {
-                            // Add and sort
+                            case TransactionVersion.OracleRequest:
+                                {
+                                    // Check the cached contract
 
-                            _orderedPool.Add(tx);
-                            _isDirty = true;
+                                    if (_lastContract?.ScriptHash != contract.ScriptHash)
+                                    {
+                                        // Reduce the memory load using the same Contract class
 
-                            // Pop one item if it's needed
+                                        _lastContract = contract;
+                                    }
 
-                            if (_asyncPool.Count <= 0)
-                            {
-                                PopTransaction();
-                            }
+                                    // If it's an OracleRequest and it's new, tell it to OracleService
+
+                                    if (_pendingOracleRequest.TryAdd(tx.Hash, new RequestItem(contract, tx)))
+                                    {
+                                        ReverifyPendingResponses(snapshot, tx.Hash);
+                                        //Self.Tell(tx);
+                                    }
+
+                                    // Process oracle - Pop one item if it's needed
+
+                                    if (_queue.TryAdd(tx.Hash, tx) && _asyncPool.Count <= 0)
+                                    {
+                                        PopTransaction();
+                                    }
+
+                                    break;
+                                }
+                            case TransactionVersion.OracleResponse:
+                                {
+                                    var hashes = tx.GetScriptHashesForVerifying(snapshot);
+
+                                    if (hashes.Length != 1 || hashes[0] != contract.ScriptHash)
+                                    {
+                                        break;
+                                    }
+
+                                    // We should receive only this transactions P2P, never from OracleService
+
+                                    if (tx.VerifyWitnesses(snapshot, 200_000_000))
+                                    {
+                                        ReverifyPendingResponses(snapshot, tx.OracleRequestTx);
+                                    }
+
+                                    // TODO: Send it to mempool?
+
+                                    break;
+                                }
                         }
                         break;
                     }
@@ -142,19 +314,8 @@ namespace Neo.Oracle
         /// </summary>
         private void PopTransaction()
         {
-            if (_orderedPool.Count > 0)
+            if (_queue.TryPop(out var entry))
             {
-                if (_isDirty && _orderedPool.Count > 1)
-                {
-                    // It will require a sort before pop the item
-
-                    _orderedPool.Sort((a, b) => b.FeePerByte.CompareTo(a.FeePerByte));
-                    _isDirty = false;
-                }
-
-                var entry = _orderedPool[0];
-                _orderedPool.RemoveAt(0);
-
                 _asyncPool.Add(entry);
             }
         }
@@ -176,12 +337,8 @@ namespace Neo.Oracle
                 {
                     foreach (var tx in _asyncPool.GetConsumingEnumerable(_cancel.Token))
                     {
-                        ProcessTransaction(tx);
-
-                        lock (_orderedPool)
-                        {
-                            PopTransaction();
-                        }
+                        ProcessRequestTransaction(tx);
+                        PopTransaction();
                     }
                 },
                 _cancel.Token);
@@ -213,7 +370,9 @@ namespace Neo.Oracle
 
             // Clean queue
 
-            _orderedPool.Clear();
+            _queue.Clear();
+            _pendingOracleRequest.Clear();
+            _pendingOracleResponses.Clear();
 
             while (_asyncPool.Count > 0)
             {
@@ -222,52 +381,51 @@ namespace Neo.Oracle
         }
 
         /// <summary>
-        /// Process transaction
+        /// Process request transaction
         /// </summary>
         /// <param name="tx">Transaction</param>
-        private void ProcessTransaction(Transaction tx)
+        private void ProcessRequestTransaction(Transaction tx)
         {
             var oracle = new OracleExecutionCache(Process);
             using var snapshot = Blockchain.Singleton.GetSnapshot();
             using var engine = new ApplicationEngine(TriggerType.Application, tx, snapshot, tx.SystemFee, false, oracle);
             engine.LoadScript(tx.Script);
 
-            if (engine.Execute() == VMState.HALT)
+            if (engine.Execute() != VMState.HALT)
             {
-                // Create deterministic oracle response
-
-                var responseTx = CreateResponseTransaction(snapshot, oracle, tx);
-
-                foreach (var keyPair in _accounts)
-                {
-                    // Sign the transaction
-
-                    Sign(responseTx, keyPair, out var signature);
-
-                    // Create the payload
-
-                    var response = new OraclePayload()
-                    {
-                        OraclePub = keyPair.PublicKey,
-                        OracleSignature = new OracleResponseSignature()
-                        {
-                            OracleExecutionCacheHash = oracle.Hash,
-                            Signature = signature,
-                            TransactionRequestHash = tx.Hash
-                        }
-                    };
-
-                    if (Pool.TryAddOracleResponse(snapshot, response, responseTx))
-                    {
-                        // Send my signature by P2P
-
-                        _localNode.Tell(response);
-                    }
-                }
+                // TODO: If the request TX will FAULT?
+                // oracle.Clear();
             }
-            else
+
+            // Create deterministic oracle response
+
+            var responseTx = CreateResponseTransaction(snapshot, oracle, tx);
+
+            foreach (var keyPair in _accounts)
             {
-                // TODO: Send something or not
+                // Sign the transaction
+
+                var signature = responseTx.Sign(keyPair);
+
+                // Create the payload
+
+                var response = new OraclePayload()
+                {
+                    OraclePub = keyPair.PublicKey,
+                    OracleSignature = new OracleResponseSignature()
+                    {
+                        OracleExecutionCacheHash = oracle.Hash,
+                        Signature = signature,
+                        TransactionRequestHash = tx.Hash
+                    }
+                };
+
+                if (TryAddOracleResponse(snapshot, response, responseTx))
+                {
+                    // Send my signature by P2P
+
+                    _localNode.Tell(response);
+                }
             }
         }
 
@@ -284,7 +442,7 @@ namespace Neo.Oracle
             var sender = NativeContract.Oracle.GetOracleMultiSigAddress(snapshot);
             using ScriptBuilder script = new ScriptBuilder();
 
-            script.EmitAppCall(NativeContract.Oracle.Hash, "setOracleResponse", requestTx.Hash, oracle.ToArray());
+            script.EmitAppCall(NativeContract.Oracle.Hash, "setOracleResponse", requestTx.Hash, IO.Helper.ToArray(oracle));
 
             return new Transaction()
             {
@@ -309,17 +467,106 @@ namespace Neo.Oracle
             };
         }
 
-        /// <summary>
-        /// Sign
-        /// </summary>
-        /// <param name="tx">Transaction</param>
-        /// <param name="account">Account</param>
-        /// <param name="signature">Signature</param>
-        private void Sign(Transaction tx, KeyPair account, out byte[] signature)
-        {
-            // Sign the transaction and extract the signature of this oracle
+        #region Sorts
 
-            signature = new byte[0];
+        private int SortRequest(KeyValuePair<UInt256, RequestItem> a, KeyValuePair<UInt256, RequestItem> b)
+        {
+            return a.Value.CompareTo(b.Value);
+        }
+
+        private int SortEnqueuedRequest(KeyValuePair<UInt256, Transaction> a, KeyValuePair<UInt256, Transaction> b)
+        {
+            return a.Value.FeePerByte.CompareTo(b.Value.FeePerByte);
+        }
+
+        private int SortResponse(KeyValuePair<UInt256, ResponseCollection> a, KeyValuePair<UInt256, ResponseCollection> b)
+        {
+            return a.Value.Timestamp.CompareTo(b.Value.Timestamp);
+        }
+
+        #endregion
+
+        /// <summary>
+        /// Try add oracle response payload
+        /// </summary>
+        /// <param name="snapshot">Snapshot</param>
+        /// <param name="oracle">Oracle</param>
+        /// <param name="responseTx">Response TX (from OracleService)</param>
+        /// <returns>True if it was added</returns>
+        internal bool TryAddOracleResponse(StoreView snapshot, OraclePayload oracle, Transaction responseTx)
+        {
+            return TryAddOracleResponse(snapshot, new ResponseItem(oracle, responseTx));
+        }
+
+        /// <summary>
+        /// Try add oracle response payload
+        /// </summary>
+        /// <param name="snapshot">Snapshot</param>
+        /// <param name="response">Response</param>
+        /// <returns>True if it was added</returns>
+        private bool TryAddOracleResponse(StoreView snapshot, ResponseItem response)
+        {
+            if (!response.Msg.Verify(snapshot))
+            {
+                return false;
+            }
+
+            // Find the request tx
+
+            if (_pendingOracleRequest.TryGetValue(response.Data.TransactionRequestHash, out var request))
+            {
+                // Append the signature if it's possible
+
+                if (request.AddSinature(response))
+                {
+                    if (request.IsCompleted)
+                    {
+                        // Done! Send to mem pool
+
+                        MemPool.TryAdd(request.ResponseTransaction.Hash, request.ResponseTransaction);
+                        MemPool.TryAdd(request.RequestTransaction.Hash, request.RequestTransaction);
+
+                        _pendingOracleRequest.TryRemove(response.Data.TransactionRequestHash, out _);
+                    }
+
+                    return true;
+                }
+            }
+
+            // Save this payload for check it later
+
+            if (_pendingOracleResponses.TryGetValue(response.Data.TransactionRequestHash, out var collection))
+            {
+                return collection.Add(response);
+            }
+            else
+            {
+                // TODO: This could not be thread-safe (lock?)
+
+                return _pendingOracleResponses.TryAdd(response.Data.TransactionRequestHash, new ResponseCollection(response));
+            }
+        }
+
+        /// <summary>
+        /// Reverify pending responses
+        /// </summary>
+        /// <param name="snapshot">Snapshot</param>
+        /// <param name="requestTx">Request transaction hash</param>
+        private void ReverifyPendingResponses(StoreView snapshot, UInt256 requestTx)
+        {
+            // If the response is pending, we should process it now
+
+            if (!_pendingOracleResponses.TryRemove(requestTx, out var collection))
+            {
+                return;
+            }
+
+            // Order by Transaction
+
+            foreach (var entry in collection.OrderByDescending(a => a.ResponseTx != null ? 1 : 0))
+            {
+                TryAddOracleResponse(snapshot, entry);
+            }
         }
 
         #region Public Static methods
